Update mixins.py and wms_proxy.py
